Sigil Twigil Distinctness The great STD debacle of 2015. Jokes aside, this is an attempt at doing something a bit different with highlighting of Perl6: only highlight the sigil and twigil (each a bit differently as well) and leave the variable name alone.
In some ways, I do think it's interesting.  Personally, I think it actually
reduces visual noise when reading Perl6. However, others have said that
it actually contributes to it.  After some soul-searching, I did finally
decide that especially with the `%` character (hashes and modulo), it
adds obfuscation.  So I am going to let it linger in this branch and move
on.